Added error handling and full authorization

+++ backend/routes/posts.js
@@ -33,16 +33,20 @@ router.post('', checkAuth ,multer({storage: storage}).single("image"), (req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then((createdPost) => {
-    res.status(201).json({
+    return res.status(201).json({
       message: 'Post Added Successfully',
       post: {
         ...createdPost,
         id: createdPost._id
       }
     });
+  })
+  .catch(error => {
+    return res.status(500).json({ message: "Creating a Post Failed!"});
   });
 });
 
@@ -56,26 +60,43 @@ router.put('/:id', checkAuth, multer({storage: storage}).single("image"), (req, 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  console.log(post);
-  Post.updateOne({_id: req.params.id}, post)
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post)
     .then((postData) => {
-      res.status(200).json({
-        message: 'Update Successful',
-        updatedPost: post
-      })
+      if(postData.nModified > 0) {
+        return res.status(200).json({
+          message: 'Update Successful',
+          updatedPost: post
+        });
+      } else {
+        return res.status(401).json({
+          message: 'Not authorized!'
+        });
+      }
     })
+    .catch(error => {
+      return res.status(500).json({ message:"Couldn't update Post!"});
+    });
 });
 
 router.delete('/:id', checkAuth, (req, res, next) =>{
-  Post.deleteOne({ _id: req.params.id })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then((result) => {
-      console.log(result);
+      if(result.n > 0) {
+        return res.status(200).json({
+          message: 'Deletion Successful'
+        });
+      } else {
+        return res.status(401).json({
+          message: 'Not authorized!'
+        });
+      }
+    })
+    .catch(error => {
+      return res.status(500).json({ message:"Deleting Post Failed!"});
     });
-  res.status(200).json({
-    message: 'Post Deleted'
-  })
 });
 
 router.get('', (req, res, next) => {
@@ -95,21 +116,27 @@ router.get('', (req, res, next) => {
       return Post.countDocuments();
     })
     .then((count) => {
-      res.status(200).json({
+      return res.status(200).json({
         message: 'Posts Fetched Successfully',
         posts: fetchedPosts,
         maxPosts: count
       })
+    })
+    .catch(error => {
+      return res.status(500).json({ message:"Fetcing Posts Failed!"});
     });
 });
 
 router.get('/:id', (req, res, next) => {
   Post.findById(req.params.id).then((postData) => {
     if(postData) {
-      res.status(200).json({ message: 'Post fetched successfully', post: postData })
+      return res.status(200).json({ message: 'Post fetched successfully', post: postData })
     } else {
-      res.status(404).json({ message: 'Post not found' });
+      return res.status(404).json({ message: 'Post not found' });
     }
+  })
+  .catch(error => {
+    return res.status(500).json({ message:"Fetcing Post Failed!"});
   });
 });
